cleans up writable data source config
