+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/WritableDataSourceConfiguration.java
@@ -18,6 +18,67 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import javax.sql.DataSource;
 
 @Configuration
+@EnableTransactionManagement
 public class WritableDataSourceConfiguration {
+
+    /**
+     * DataSourceProperties
+     *
+     * @return DataSourceProperties
+     */
+    @Bean(name = "writableDataSourceElementsProperties")
+    @ConfigurationProperties("datasources.reporting_rw")
+    @RefreshScope
+    public DataSourceElementsProperties dataSourceElementsProperties() {
+        return new DataSourceElementsProperties();
+    }
+
+    /**
+     * writable DataSourceElementsResolver
+     *
+     * @param tenantKeyResolver            key resolver
+     * @param dataSourceElementsProperties config properties
+     * @return DataSourceElementsResolver
+     */
+    @Bean(name = "writableDataSourceElementsResolver")
+    @RefreshScope
+    public DataSourceElementsResolver writableDataSourceElementsResolver(TenantKeyResolver tenantKeyResolver,
+                                                                         @Qualifier("writableDataSourceElementsProperties") DataSourceElementsProperties dataSourceElementsProperties) {
+        return new DataSourceElementsResolver(tenantKeyResolver, dataSourceElementsProperties);
+    }
+
+    /**
+     * Data source for the data warehouse.
+     *
+     * @param dataSourceElementsResolver reporting DataSourceElementsResolver
+     * @param tenantKeyResolver          reporting tenant key resolver
+     * @return DataSource
+     */
+    @Bean(name = "writableDataSource")
+    @RefreshScope
+    public DataSource writableDataSource(@Qualifier("writableDataSourceElementsResolver") DataSourceElements dataSourceElementsResolver,
+                                         TenantKeyResolver tenantKeyResolver) {
+        return new TenantDynamicRoutingDataSource(dataSourceElementsResolver, tenantKeyResolver);
+    }
+
+    /**
+     * The jdbc template used for reading from the data warehouse.
+     *
+     * @return DataSource
+     */
+    @Bean
+    public NamedParameterJdbcTemplate writableJdbcTemplate(
+            @Qualifier("writableDataSource") final DataSource dataSource) {
+        return new NamedParameterJdbcTemplate(dataSource);
+    }
+
+    /**
+     * @return The reporting transaction manager
+     */
+    @Bean
+    public PlatformTransactionManager writableDatasourceTransactionManager(
+            @Qualifier("writableDataSource") final DataSource dataSource) {
+        return new DataSourceTransactionManager(dataSource);
+    }
 
 }
